Create WeekOneTakeHome1
Post a java program to Github that utilizes the concepts of inheritance, encapsulation, and polymorphism. 

Use code comments to signify the instances where you are using inheritance, encapsulation, and polymorphism. 

Submit link to Github on Canvas in the assignments section. 


In this case the superclass is a graded activity, and the subclass is an essay assignment. 
The essay inherits the grading system from graded activity, but adds a rubric, which makes the overall grade the average of 4 grading criteria (polymorphism). It uses encapsulation by making the parameters private and only allowing access through the public methods available.